flytt barnehagebarn-interface til Barnehagebarn og rename -> BarnehagebarnPaginerbar
