+++ src/main/kotlin/no/nav/familie/ks/sak/api/dto/BarnehagebarnRequestParams.kt
@@ -1,5 +1,7 @@
 package no.nav.familie.ks.sak.api.dto
 
+import org.springframework.data.domain.Sort
+
 data class BarnehagebarnRequestParams(
     val ident: String?,
     val kommuneNavn: String?,
@@ -9,3 +11,19 @@ data class BarnehagebarnRequestParams(
     val sortBy: String = "kommuneNavn",
     val sortAsc: Boolean = false,
 )
+
+fun BarnehagebarnRequestParams.toSort() =
+    if (sortAsc) {
+        Sort.by(getCorrectSortBy(sortBy)).ascending()
+    } else {
+        Sort.by(getCorrectSortBy(sortBy)).descending()
+    }
+
+private fun getCorrectSortBy(sortBy: String): String =
+    when (sortBy.lowercase()) {
+        "endrettidspunkt" -> "endretTid"
+        "kommunenavn" -> "kommune_navn"
+        "kommunenr" -> "kommune_nr"
+        "antalltimeribarnehage" -> "antall_timer_i_barnehage"
+        else -> sortBy
+    }
